+++ src/main/java/net/silentchaos512/gear/data/client/ModBlockStateProvider.java
@@ -3,12 +3,14 @@ package net.silentchaos512.gear.data.client;
 import net.minecraft.block.FenceBlock;
 import net.minecraft.data.DataGenerator;
 import net.minecraft.state.properties.BlockStateProperties;
+import net.minecraft.util.Direction;
 import net.minecraft.util.ResourceLocation;
 import net.minecraftforge.client.model.generators.*;
 import net.minecraftforge.common.data.ExistingFileHelper;
 import net.silentchaos512.gear.SilentGear;
 import net.silentchaos512.gear.block.ModCropBlock;
 import net.silentchaos512.gear.init.ModBlocks;
+import net.silentchaos512.lib.block.IBlockProvider;
 import net.silentchaos512.lib.util.NameUtils;
 
 import javax.annotation.Nonnull;
@@ -28,6 +30,7 @@ public class ModBlockStateProvider extends BlockStateProvider {
 
     @Override
     protected void registerStatesAndModels() {
+        // Ores and storage blocks
         simpleBlock(ModBlocks.BORT_ORE.get());
         simpleBlock(ModBlocks.CRIMSON_IRON_ORE.get());
         simpleBlock(ModBlocks.AZURE_SILVER_ORE.get());
@@ -39,6 +42,7 @@ public class ModBlockStateProvider extends BlockStateProvider {
         simpleBlock(ModBlocks.AZURE_ELECTRUM_BLOCK.get());
         simpleBlock(ModBlocks.TYRIAN_STEEL_BLOCK.get());
 
+        // Netherwood
         simpleBlock(ModBlocks.NETHERWOOD_CHARCOAL_BLOCK.get());
         simpleBlock(ModBlocks.NETHERWOOD_PLANKS.get());
         simpleBlock(ModBlocks.NETHERWOOD_LEAVES.get());
@@ -55,6 +59,7 @@ public class ModBlockStateProvider extends BlockStateProvider {
         doorBlock(ModBlocks.NETHERWOOD_DOOR.get(), modLoc("block/netherwood_door_bottom"), modLoc("block/netherwood_door_top"));
         trapdoorBlock(ModBlocks.NETHERWOOD_TRAPDOOR.get(), modLoc("block/netherwood_trapdoor"), true);
 
+        // Fluffy blocks
         simpleBlock(ModBlocks.WHITE_FLUFFY_BLOCK.get());
         simpleBlock(ModBlocks.ORANGE_FLUFFY_BLOCK.get());
         simpleBlock(ModBlocks.MAGENTA_FLUFFY_BLOCK.get());
@@ -72,6 +77,7 @@ public class ModBlockStateProvider extends BlockStateProvider {
         simpleBlock(ModBlocks.RED_FLUFFY_BLOCK.get());
         simpleBlock(ModBlocks.BLACK_FLUFFY_BLOCK.get());
 
+        // Oddballs
         simpleBlock(ModBlocks.PHANTOM_LIGHT.get(), models().cubeAll("phantom_light", modLoc("item/blank")));
         simpleBlock(ModBlocks.POTTED_NETHERWOOD_SAPLING.get(), models()
                 .withExistingParent("potted_netherwood_sapling", "block/flower_pot_cross")
@@ -83,66 +89,26 @@ public class ModBlockStateProvider extends BlockStateProvider {
                         .rotationY((int) state.get(BlockStateProperties.HORIZONTAL_FACING).getHorizontalAngle() + 90)
                         .build());
 
-        {
-            ResourceLocation bottom = modLoc("block/gear_smithing_table_bottom");
-            ResourceLocation front = modLoc("block/gear_smithing_table_front");
-            ResourceLocation side = modLoc("block/gear_smithing_table_side");
-            ResourceLocation top = modLoc("block/gear_smithing_table_top");
-            simpleBlock(ModBlocks.GEAR_SMITHING_TABLE.get(),
-                    models().cube("gear_smithing_table", bottom, top, front, front, side, side)
-                            .texture("particle", front));
-        }
+        // Crafters and Machines
+        simpleBlock(ModBlocks.GEAR_SMITHING_TABLE.get(), getExistingModel("gear_smithing_table"));
+        simpleBlock(ModBlocks.STARLIGHT_CHARGER.get(), getExistingModel("starlight_charger"));
+        orientableMachineBlock(ModBlocks.MATERIAL_GRADER, "material_grader");
+        orientableMachineBlock(ModBlocks.METAL_PRESS, "metal_press");
+        orientableMachineBlock(ModBlocks.SALVAGER, "salvager");
 
         // Compounders
-        {
-            BlockModelBuilder offModel = models().orientableWithBottom("metal_alloyer",
-                    modLoc("block/metal_alloyer_side"),
-                    modLoc("block/metal_alloyer_front"),
-                    modLoc("block/metal_alloyer_bottom"),
-                    modLoc("block/metal_alloyer_top"));
-            BlockModelBuilder onModel = models().orientableWithBottom("metal_alloyer_on",
-                    modLoc("block/metal_alloyer_side"),
-                    modLoc("block/metal_alloyer_front_on"),
-                    modLoc("block/metal_alloyer_bottom"),
-                    modLoc("block/metal_alloyer_top"));
-            horizontalFaceBlock(ModBlocks.METAL_ALLOYER.get(), state ->
-                    state.get(BlockStateProperties.LIT) ? onModel : offModel);
-        }
-        {
-            BlockModelBuilder offModel = models().cubeTop("recrystallizer",
-                    modLoc("block/recrystallizer_side"),
-                    modLoc("block/recrystallizer_top"));
-            BlockModelBuilder onModel = models().cubeTop("recrystallizer_on",
-                    modLoc("block/recrystallizer_side"),
-                    modLoc("block/recrystallizer_top_on"));
-            getVariantBuilder(ModBlocks.RECRYSTALLIZER.get()).forAllStates(state -> {
-                return ConfiguredModel.builder()
-                        .modelFile(state.get(BlockStateProperties.LIT) ? onModel : offModel)
-                        .build();
-            });
-        }
-//        {
-//            BlockModelBuilder offModel = models().cubeTop("refabricator",
-//                    modLoc("block/refabricator_side"),
-//                    modLoc("block/refabricator_top"));
-//            BlockModelBuilder onModel = models().cubeTop("refabricator_on",
-//                    modLoc("block/refabricator_side"),
-//                    modLoc("block/refabricator_top_on"));
-//            getVariantBuilder(ModBlocks.REFABRICATOR.get()).forAllStates(state -> {
-//                return ConfiguredModel.builder()
-//                        .modelFile(state.get(BlockStateProperties.LIT) ? onModel : offModel)
-//                        .build();
-//            });
-//        }
-        simpleBlock(ModBlocks.REFABRICATOR.get()); // temp
+        orientableMachineBlock(ModBlocks.METAL_ALLOYER, "metal_alloyer");
+        orientableMachineBlock(ModBlocks.RECRYSTALLIZER, "recrystallizer");
+        orientableMachineBlock(ModBlocks.REFABRICATOR, "refabricator");
 
+        // Plants
         getVariantBuilder(ModBlocks.FLAX_PLANT.get()).forAllStates(state -> {
             int i = cropAgeToIndex(state.get(ModCropBlock.AGE));
             return ConfiguredModel.builder()
                     .modelFile(models().crop("flax_plant" + i, modLoc("block/flax_plant" + i)))
                     .build();
         });
-        simpleBlock(ModBlocks.WILD_FLAX_PLANT.get(), models().crop("wild_flax_plant", SilentGear.getId("block/flax_plant3")));
+        simpleBlock(ModBlocks.WILD_FLAX_PLANT.get(), models().crop("wild_flax_plant", modLoc("block/flax_plant3")));
 
         getVariantBuilder(ModBlocks.FLUFFY_PLANT.get()).forAllStates(state -> {
             int i = cropAgeToIndex(state.get(ModCropBlock.AGE));
@@ -150,7 +116,24 @@ public class ModBlockStateProvider extends BlockStateProvider {
                     .modelFile(models().crop("fluffy_plant" + i, modLoc("block/fluffy_plant" + i)))
                     .build();
         });
-        simpleBlock(ModBlocks.WILD_FLUFFY_PLANT.get(), models().crop("wild_fluffy_plant", SilentGear.getId("block/fluffy_plant3")));
+        simpleBlock(ModBlocks.WILD_FLUFFY_PLANT.get(), models().crop("wild_fluffy_plant", modLoc("block/fluffy_plant3")));
+    }
+
+    private ModelFile.ExistingModelFile getExistingModel(String blockName) {
+        return models().getExistingFile(modLoc(blockName));
+    }
+
+    private void orientableMachineBlock(IBlockProvider block, String name) {
+        ModelFile.ExistingModelFile offModel = getExistingModel(name);
+        ModelFile.ExistingModelFile onModel = getExistingModel(name + "_on");
+        getVariantBuilder(block.asBlock()).forAllStates(state -> {
+            Direction facing = state.get(BlockStateProperties.HORIZONTAL_FACING);
+            boolean lit = state.get(BlockStateProperties.LIT);
+            return ConfiguredModel.builder()
+                    .modelFile(lit ? onModel : offModel)
+                    .rotationY((int) facing.getOpposite().getHorizontalAngle())
+                    .build();
+        });
     }
 
     public ModelBuilder<BlockModelBuilder> wallTorch(String name, ResourceLocation torch) {